msm: board-9615: Configure PM8018 LDO 3 to be always-on
Modify the configuration of PMIC PM8018 LDO 3 to have
always_on=1.  This will ensure that Linux will never send a
disable request to the RPM for LDO 3.  Keeping L3 enabled in this
fashion is required because the modem SR2 PLL is powered by L3
and the modem cannot easily send a request to the RPM to keep
L3 enabled when it is needed.  Instead, the Apps processor is
proxy voting for the modem processor for L3.

Change-Id: I048850e80c0cb622d8afd9250c81b8709e01e68d
CRs-Fixed: 347257
Signed-off-by: David Collins <collinsd@codeaurora.org>