ASoC: msm: update BTSCO rate put function.
The value for the amix command with the numeric strings is
the actual value not the order of the string in the array
of text so the put function has to be updated to handle it.
For example,  the following amix comand:
amix 'Internal BTSCO SampleRatee' 16000'
The value passed from the alsa-lib is not 2, i.e., the 2nd
paramater in the command string:
static const char *btsco_rate_text[] = {"8000", "16000"};
Instead, it is 16000.

CRs-fixed: 364832

Signed-off-by: Kuirong Wang <kuirongw@codeaurora.org>
(cherry picked from commit fd8e0b303ba46207f24a3738f8859e84044833fd)

Change-Id: Id3860564e45aa12bf95774fc523ffee5bf4aa487
Signed-off-by: Ram Kumar Chakravarthy Chebathini <rcheba@codeaurora.org>