board: msm7x27a: Correct the operating voltage range.
As per the PMIC team right operating voltage range is 3.2 to 4.2V.
So correcting the voltage range to this. The Android framework pops
up low battery warning at 14% of the remaining baterry capacity.
With this new range 14% would be at 3.34V. So fill this voltage in
fail_safe_voltage in the board file so that target wakes up incase
it reaches this voltage when its asleep.

CRs-Fixed: 366506
Signed-off-by: Krishna Vanka <kvanka@codeaurora.org>
(cherry picked from commit 9a265e1b6b713138236e70241cbeace7b70f0878)

Change-Id: Id6cf2d9a7a7364516a2d9c84185819309e6a821d
Signed-off-by: Sridhar Gujje <sgujje@codeaurora.org>