msm: ipc: Add support to load and unload the subsystems
When all the user-space modules communicating with a subsystem get
killed, the kernel releases the resources consumed by the modules. One
of the resources is SMD channel. Hence the APPS processor closes its end
of the SMD channel. The concerned drivers like SMD_PKT, SMD_TTY
trigger unloading the subsystem with which the module has been
communicating. When the user-space modules come up again, the concerned
subsystem gets reloaded. As part of boot-up, the subsystems reset their
end of the SMD channels.

IPC Router, being a kernel driver, is unaware of this unloading and
reloading of the subsystem. Hence the APPs side state of the SMD
channel, used by IPC Router, is left undisturbed but the remote side is
reset. This leaves the concerned SMD channel in stale state. The
subsystem while handling the stale SMD channel crashes. Hence the IPC
Router will cast a vote to PIL to ensure that the remote subsystem is
loaded as long as it is using the SMD channel.

CRs-Fixed: 359773
Signed-off-by: Karthikeyan Ramasubramanian <kramasub@codeaurora.org>
(cherry picked from commit c29692cb2421ce8317da15ad39301d328b226842)

Change-Id: Ifa00eb8c5a7b9862ac969477a84fa9e90e4d5487
Signed-off-by: Ram Kumar Chakravarthy Chebathini <rcheba@codeaurora.org>