msm: kgsl: For IOMMU only decide to flush when memory is unmapped
In case of IOMMU memory needs to only be flushed when it is
unmapped and not mappped. Also, when an application is running
the number of unmaps would in most cases be less than number of
maps, so this way the number of flushes wil be fewer.

CRs-fixed: 380834
Signed-off-by: Shubhraprakash Das <sadas@codeaurora.org>
(cherry picked from commit 0c811269372d83d18538c74ace76554ac77ce933)

Change-Id: I7ae96adab7b336221fc6bd4e6e085180122c79e3
Signed-off-by: Sridhar Gujje <sgujje@codeaurora.org>