power: bms/ccadc: move ccadc calibration to ccadc driver
There is a dependency for the CCADC calibration work on the BMS
driver. This prevents some devices that do not use the BMS
to utilize the CCADC.

Remove this dependency by moving the ccadc calibration
work to the CCADC driver to avoid this.

Signed-off-by: David Keitel <dkeitel@codeaurora.org>
(cherry picked from commit 3c3788281e7979d1cb36f9424fe9c5c4806d916f)

Change-Id: I72ed6f6f6c36eda5e1aed040a25ab754d11e3d0e
Signed-off-by: Shruthi Krishna <skrish@codeaurora.org>