msm_fb: display: HPD rework for cleaner approach
Currently HPD is turned on and off with HDMI turn on and off
and turned on again after HDMI off so as to detect the cable
plugin.
With this change, HPD will be turned on when HDMI drvier is
initialized and turned off and on with suspend - resume. The
HPD handling during HDMI turn on and off has been removed.

CRs-Fixed: 368865
Signed-off-by: Aravind Venkateswaran <aravindh@codeaurora.org>
Signed-off-by: Ajay Singh Parmar <aparmar@codeaurora.org>
(cherry picked from commit 56672d222b15602826d0200a6baee5ca02a450c6)

Change-Id: I78fd38a0316a5d3c178e9b1771528dab9c00b44d
Signed-off-by: Shruthi Krishna <skrish@codeaurora.org>