msm: subsystem_restart: Convert subsystem_list_lock to a spinlock.
Use a spinlock instead of a mutex for subsystem_list_lock,
since this lock may be required in both interrupt and
pre-emptive context. A small list is traversed when this
lock is held, and therefore the penalty of holding this
lock is minimal.

Change-Id: I886747dc77d539c4d572e832fd676314f8e66b9f
Signed-off-by: Vikram Mulukutla <markivx@codeaurora.org>
(cherry picked from commit 705e68969220f568ebea9376c50e80e18e58d5b3)