cpufreq: synchronous cpufreq support for apq8064
Enable the synchronous cpufreq support for 8064. If the parameter
msm_force_sync_cpufreq is set, the related cpus policy will be enabled,
and the frequency of all the online cpus will be set for every frequency
change request. The other flag msm_sync_cpufreq is to guard the cpus like
8625 which are already synchronous. Also, When a new cpu is added the
frequency of the new cpu will be aligned to the core0 frequency.

Change-Id: If42ab057c4d09253879dc8bf162391898c8cea1f
Signed-off-by: Narayanan Gopalakrishnan <nargop@codeaurora.org>