msm: kgsl: Improve detection of a type3 CP packet
Currently we just check the most significant two bits of a dword
in the ringbuffer or IB are set to decide if the data is a type3
packet header.  This has proven to be problematic, as it turns out
that the top two bits are often set in nominal data.  Enhance
the type3 packet check to verify that various reserved bits in
the dword are 0 as expected.

Signed-off-by: Jordan Crouse <jcrouse@codeaurora.org>
(cherry picked from commit 2012c8d705b7ae6ddacc2a4dc0eb6d195f0c1469)

Change-Id: I844984a66509a7e81a2dfcb0d00b98654f922c22
Signed-off-by: Shruthi Krishna <skrish@codeaurora.org>