Bluetooth: Store the MTU if configuration of other params fail
Store the MTU to use the proper MTU requested by remote if the
configuration fails for other options. If the remote does not
send the MTU in the next configuration request,
we need to use that value.

CRs-fixed: 373233
Signed-off-by: Mallikarjuna GB <gbmalli@codeaurora.org>
(cherry picked from commit 647b004ede4061d336536f68aece062d5c20b984)

Change-Id: I291cca1e5bb6a97353cef3b45128ac02355bf3d6
Signed-off-by: Sridhar Gujje <sgujje@codeaurora.org>