msm_fb: MHL: Board file changes for Integrated MHL
8x30 FLUID device has integrated solution for MHL.
Power and GPIO related changes for MHL feature
is added. This is the first solution
where 8334 chipset is integrated as the MHL tx. The power
up for rails and MUX selection is specific to 8x30, since
8960 doesn't yet have this solution. This change also
includes the change to move the power and GPIO functions
to the driver from the board file.

Signed-off-by: Manoj Rao <manojraj@codeaurora.org>
Signed-off-by: Pavankumar Kondeti <pkondeti@codeaurora.org>
(cherry picked from commit c6d904ce082c15a23234c25beb789f75490589d4)

Change-Id: Idcec5e86d692c41f9bb1d3684185ecc3314f6c58
Signed-off-by: Ram Kumar Chakravarthy Chebathini <rcheba@codeaurora.org>