msm: clock-rpm: Fix branch clocks so that they can be enabled
RPM branch clocks do not have settable rates. The DEFINE_CLK_RPM_BRANCH()
macro assigns hard-coded rates at compile-time which we rely on being
non-zero for rpm_clk_enable() to work.

The current handoff code queries the state of RPM clocks and assigns the
resulting rate (which can be zero) to the clock. In the case of RPM branch
clocks this is bad, since it overwrites the rate set at compile-time. This
results in branch clocks failing to enable if the queried rate was zero.

Fix this by not querying the rate RPM for branch clocks so that the rates
set at compile-time are preserved.

Signed-off-by: Matt Wagantall <mattw@codeaurora.org>
(cherry picked from commit 215de5dcbdc3e5ee3504bbdfe5c577ce66a2851f)

Change-Id: I672b9953678b806591425dfd196f7d4dd045ad9b
Signed-off-by: Swetha Basineni <sbasin@codeaurora.org>