msm: tty: update receive room just before writing data to the ldisc
There is a corner case in the tty driver where the value of
tty->receive_room is not being updated before writing data to
the line discipline. When this happens, data is lost because
of failure to re-submit any remaining data to the ldisc.

This fix is dependent on a new tty flag that is set only when
the tty driver is used for efs sync betweem the mdm modem and
the applications processor.

CRs-Fixed: 358868
Signed-off-by: Joel King <joelking@codeaurora.org>
(cherry picked from commit 913565a83d74fe1da0914e8c355345bc616131c4)

Change-Id: Ieda5cb13b455aded59c6dc8320e4026f8a19cb1e
Signed-off-by: Sridhar Gujje <sgujje@codeaurora.org>