msm: clock: Remove some clk_prepare/unprepare() warnings
Remove the warnings that the clk_enable() and clk_disable() APIs
print when drivers attempt to enable or disable clocks that are
not prepared. These warnings are currently harmless, but greatly
contribute to noise in the kernel logs.

This is safe because use of clk_prepare() and clk_unprepare() is
currently optional. They are placeholder APIs that do not
currently perform any operations. Their use will eventually be
required in future kernels, but never on this kernel branch.

Change-Id: I4239842c6f77b844e5dd8460229b80d8c51e99bb
Signed-off-by: Matt Wagantall <mattw@codeaurora.org>