msm: bam_dmux: update SSR handling to reflect disconnect ack on by default
When subsystem restart of the modem is triggered, the SSR handling function
disables the disconnect ack feature.  Later, when the modem reboots, it
expects the disconnect ack feature to be enabled by default as it was when
the modem first booted.  This causes a problem when the modem times out
waiting for the disconnect ack and asserts.

Fix the issue by changing the SSR code in bam_dmux to reenable the
disconnect ack feature instead of disabling it.

CRs-Fixed: 377493

Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>
(cherry picked from commit f62029d1c902342470816ad4920b08ea805a52d6)

Change-Id: I6ce8e1099725e9037881d972e7bdddce3df18314
Signed-off-by: Dhivya Subramanian <dthiru@codeaurora.org>