msm: footswitch-8x60: Reset DSI domain along with MDP collapse/restore
Some DSI-related logic lives in MDP's collapsible GDHS domain for
8960, 8064 and 8930. Assert the reset to this domain when the reset
of MDP is reset. Use of the reset-only clocks added here assumes the
display driver is making sure the DSI clocks (not managed by the
clock driver) are enabled for the duration of the collapse and
restore sequences so that the resets propagate.

Change-Id: I0f7e6ff7b87778b798012fe3f18407f352332e46
Signed-off-by: Matt Wagantall <mattw@codeaurora.org>
Signed-off-by: Sridhar Gujje <sgujje@codeaurora.org>