msm: acpuclock-8960: Remove use of QSB as a CPU or L2 clock source
QSB's rate is tied to the Apps fabric, which means that its rate is
unpredictable. When the CPU is running at a low voltage, if the QSB
clock source is selected, it's possible that the CPU clock rate
could increase beyond the safe limit for that voltage.

Instead of selecting QSB for power-collapse and hotplug scenarios,
select an always-on source with a predictable rate (PLL8 @ 384MHz).

Change-Id: Idcff6fdfc609821ee5d4ef29e4606f9c00008f13
Signed-off-by: Matt Wagantall <mattw@codeaurora.org>