gpu/msm: don't increase the fail by blocking signals (v2)
In some error paths, if the gpu has hung, we'll spin forever without
even checking if there is a pending signal.  Which makes ctrl-c not
even work, and somehow prevents gpu recovery from kicking in.  (I
guess a lock contention between userspace trying to submit cmds and
the recovery code.)