usb: gadget: Resume usb bus on reset
When receiving reset interrupt from the usb controller, check if the bus
is suspended. If so then first resume the bus and only then reset
the usb controller.

This also fixes the CRs below, in a case of race condition when reset
interrupt is happening right after suspend interrupt. In this case
there might be a race in which the msm_otg will execute the suspend
event in the work queue after the reset interrupt finished. In this case
the bus will be suspended after the reset and no further event (resume/reset)
will arrive from host, keeping the device in LPM forever.

CRs-Fixed: 381806, 386464
Change-Id: I40fd4b245d0f7776602412444ef124dac2b46306
Signed-off-by: Ido Shayevitz <idos@codeaurora.org>