thermal: tsens8960: Fix delay in the ISR
There are more number of sensors on APQ8064 and MSM8930
whose status registers are to be read and checked
against enabled sensors if thresholds interrupts are
crossed. This causes delay in processing the ISR before
it is returned causing other timer events to be delayed.

Fix this by performing the status register reads for
the sensors in a seperate workqueue.

CRs-Fixed: 390508
Change-Id: I7b254d0d73dfefb667e7bf9484ae704fe4ff38a8
Signed-off-by: Siddartha Mohanadoss <smohanad@codeaurora.org>
(cherry picked from commit 7dc2eb0488712eb9fc2c5ebfba56c0f453f6cdce)

Signed-off-by: Dhivya Subramanian <dthiru@codeaurora.org>