ASoC: msm: Reset the pending buffer state in EOS ACK from DSP.
In the repetitive mode, after EOS of first iteration of playback
FLUSH is issued to seek the playback to begin in the second
iteration. Write call is not being issued after FLUSH as pending
buffer state is not reset with the EOS ACK from DSP. Add fix for
the same.

CRs-Fixed: 381455
Change-Id: Iea20c572f0aaaa587ddbbdce63a1519cd6500b71
Signed-off-by: Subhash Chandra Bose Naripeddy <snariped@codeaurora.org>
Signed-off-by: Aviral Gupta <aviralg@codeaurora.org>