ASoC: msm: Support send flush to DSP Multichannel PCM playback
In the use case of multi channel PCM playback with the playback
control sequence of Pause, Flush, Resume; residual data buffered in
DSP is played when resuming the playback after seek or flush.
Add support to send CMD_FLUSH to DSP to flush the buffered residual
data after Pause.

CRs-Fixed: 380473
Change-Id: I1d099d28e51858c7c22df88253cd6f83a9c3ca56
Signed-off-by: Subhash Chandra Bose Naripeddy <snariped@codeaurora.org>
Signed-off-by: Aviral Gupta <aviralg@codeaurora.org>