ASoC: msm: Fix the AUX PCM ports spurious start and stop
If the shutdown of the dai driver is called before the
prepare for the AUX PCM port then the port is closed down
and updates the incorrect aux_pcm_count.
This results in spurious start and stop of the AUX PCM ports.

To track whether the port is started or not update the
flag status_mask of the dai data.

(cherry picked from commit 04091d38160da299e1a0b10f1c695c0f889ff657)

Change-Id: Ic916200d90e42b68fd197b823ed7be7e919d82c0
CRs-Fixed: 416502
Signed-off-by: Aviral Gupta <aviralg@codeaurora.org>