ASoC: msm: Add volume control to set volume on a platform driver
The existing interface uses combinatin of k-control in routing driver
and a global structure defined in the platform driver to send volume
to DSP. Add support to use the k-control defined in platform driver
for volume control. The platform driver can add the k-control by the
volume API snd_pcm_add_volume_ctls().

Change-Id: Ia80151226f0d8d8f52670d70aea912f0cdc34642
Signed-off-by: Subhash Chandra Bose Naripeddy <snariped@codeaurora.org>