msm_fb: hdmi: Support additional CEC message mapping
This change includes additional CEC message mapping.

1. <standby> opcode = 0x36, no operand, direct addressed or broadcast

This message is used to switch all or a device into standby state and if this
message is received in active state, the HDMI driver maps it to KEY_POWER Linux
kernel key event and send it to user space. If received in standby state, do
nothing.

2. <Set Stream Path> opcode = 0x86, operand = physical addr, broadcast

This message is used to switch to a device with the given physical address. In
case this device is in suspend state, it shall wake up the device. To achieve
this, if this message is received in suspend state, this message is mapped to
KEY_POWER Linux kernel key event and sent to the user space.

3. <Play> opcode = 0x41, operand = [Play Mode], direct addressed

The two play modes - "Play Foward" and "Play Still", are mapped to KEY_PLAY and
KEY_PAUSE Linux kernel key event respectively and sent to user space, and it's
application's responsibility to configure them as a wake up event.

NOTE that system power state change using <standy> and <set stream path> must be
used with deep suspend mode (meaning that it shall be used with all USB cable
disconnected) because the functionality relies on checking suspend state which
is not available if the deep suspend mode is not used (suspend/resume functions
of each driver are not called in this case)

CRs-Fixed: 495119
Change-Id: Ia7f523e7c461d81ee3794f6955d3757e16bdaa37
Signed-off-by: Jon Lim <jonlim@codeaurora.org>