ASoC: msm: Update the hw_ptr in the CMD_RUN response
Issue:
Write is blocked after the pause call and does not recover from the resume.
During the pause command the write dones are not acknowledged and
the hw_ptr is not incremented, hence when resume is unable to unblock the
write call.
Fix:
Update the hw_ptr in the CMD_RUN response to unblock the write call.
Update the period size to accomodate the lowest period size required in HAL
Triggering the pause in Trigger stop.

CRs-Fixed: 493270 480045 488936
Change-Id: I97fdf0fbbb78f81ad6a31f6db9bf4366a5ea7642
Signed-off-by: Aviral Gupta <aviralg@codeaurora.org>