bluetooth: Update LE/ACL buffer count based on connection type
ACL buffer count is decremented as part of hci_sched_acl() though
we are performing LE operation. In the 'number of completed packets
event' we check for the connection type and increment the approrpriate
buffer count. In this case as the connection type is LE, we increment
the LE buffer count in 'num. of compl. packets event' , but decrement the
ACL buffer count in hci_sched_acl() w/o checking for connection type.
This casues the ACL buffer count to be continoulsy decremented from its
max. value of '8' to '0' and prevents sending of subsequent ACL data
packets to Controller. The appropriate buffer count(ACL/LE) has to be updated
while performing the LE/ACL operation in hci_sched_acl()and update the
correct buffer count in the number of completed packets event.

CRs-fixed: 495638

Change-Id: Idedf714822ef200a209a5b2adb42d142304a94d3
Signed-off-by: Bhakthavatsala Raghavendra <braghave@codeaurora.org>