msm: mdp: Use a waitqueue for vsync notifications
Use a waitqueue instead of completions and make sure the vsync
show event always returns the timestamp except when it is
interrupted. This simplifies the vsync wait mechanism both here
and in userspace and avoids any unnecessary locking.

Wait unconditionally in internal vsync waits.
Do not wait for timestamp to change for internal waits as the
waitqueue can be woken up from places where the vsync timestamp
doesn't change.

Change-Id: I01a243870ebdfbaeff4bf94485108f40c4e02621
Signed-off-by: Howard Pang <hpang@codeaurora.org>