ASoC: msm: Fix the race condition between the data packet writes to DSP.
Due to the distributed writes from the CMD_RUN response and the
playback_copy, there are race conditions which cause the data buffers
to be missed.
Also there can unwanted writes which can cause the target crashes.
Further there cannot be locks as the CMD_RUN response is in the
interrupt context.

Merge the data writes as the CMD_WRITE is independent on the
stream running status.

Change-Id: Ia3dc5d46745f6bf0bcbc3ec05418f188ebe5fad6
Signed-off-by: Aviral Gupta <aviralg@codeaurora.org>