Bluetooth: Add support for getting battery level
This patch adds support to retrieving battery information from HID Input
Report of Remote Device. The retrieved battery level is converted to
percentage using the battery minimum and battery maximum sent in HID
Report descriptor of Device. If there is a change in battery level
an event is sent to BlueZ user space using the management socket.

CRs-Fixed: 522195
Change-Id: I7bb301ec16f904d06e52d1626ab73f8e90e508fb
Signed-off-by: Hemant Gupta <hemantg@codeaurora.org>