hdmi_msm: Fix CEC timedout on write
Fixes timeout issue when a CEC send is exectued. It was waiting for
a threaded ISR inside the same ISR. Added a polling option with
period of HZ >> 2. The fix handles both cases: when this function
is called within the same ISR, and when this function is called
outside of the ISR.

CRs-fixed: 531377

Change-Id: Ie32a2f80f7f338c07aae8b8854d75c9d706e9ed9
Signed-off-by: Howard Pang <hpang@codeaurora.org>