msm_fb:hdmi:Fix for flicker during hotplug.
When cable is disconnected while playing audio
content, then audio engine takes around 3secs to
close and when HDMI is waiting for audio engine
to close if cable is reinserted we see the display
immediately and on handling the cable insertion
interrupt we do a core reset which causes blink
on the screen.
So, don't reset the core on cable connect/disconnect
if the video timings or mode (DVI/HDMI)are not changing.

Change-Id: I44cd0a72ea33e2c82f859fd675ba93d9f7c9f55f
Signed-off-by: Sandeep Panda <spanda@codeaurora.org>