drm: convert crtc to properties/state
Break the mutable state of a crtc out into a separate structure
and use atomic properties mechanism to set crtc attributes.  This
makes it easier to have some helpers for crtc->set_property()
and for checking for invalid params.  The idea is that individual
drivers can wrap the state struct in their own struct which adds
driver specific parameters, for easy build-up of state across
multiple set_property() calls and for easy atomic commit or roll-
back.

Conflicts:
	drivers/gpu/host1x/drm/dc.c
	drivers/staging/imx-drm/ipuv3-crtc.c