drm: Replace crtc fb with primary plane fb
Now that CRTC's have a primary plane, there's no need to track the
framebuffer in the CRTC.  Replace all references to the CRTC fb
with the primary plane's fb.

Also note that this simplifies framebuffer removal slightly; we no
longer need to scan all CRTC's and disable the ones that were using the
framebuffer since the existing loop over all planes will take care
of disabling the primary plane (and on most hardware, the CRTC by
extension).

Signed-off-by: Matt Roper <matthew.d.roper@intel.com>