WIP: drm/msm/mdp5: atomic
NOTE: currently having some problems due to hard-coded blend_setup() in
mdp5_crtc.  The end result is that we haven't *actually* stopped scanout
when mdp5_plane thinks we have, and hilarity results.  We are protected
from this pre-atomic because the crtc gets disabled first.  This should
work out much better after public planes and proper blend_setup() are in
place.