RFC: drm: add support for tiled/compressed/etc modifier in addfb2 (v1.5)
In DRM/KMS we are lacking a good way to deal with tiled/compressed
formats.  Especially in the case of dmabuf/prime buffer sharing, where
we cannot always rely on under-the-hood flags passed to driver specific
gem-create ioctl to pass around these extra flags.

The proposal is to add a per-plane format modifier.  This allows to, if
necessary, use different tiling patters for sub-sampled planes, etc.
The format modifiers are added at the end of the ioctl struct, so for
legacy userspace it will be zero padded.

TODO how best to deal with assignment of modifier token values?  The
rough idea was to namespace things with an 8bit vendor-id, and then
beyond that it is treated as an opaque value.  But that was a relatively
arbitrary choice.  There are cases where same tiling pattern and/or
compression is supported by various different vendors.  So we should
standardize to use the vendor-id and value of the first one who
documents the format?

v1: original
v1.5: increase modifier to 64b

TODO: add cap to enable, add checks in existing drivers to return error
for !=0 modifier... maybe add way to query this, but otoh we probably
need to query modifiers per format.  So new ioctl?  Maybe it is enough
to start by only caring about driver-specific userspace?

Signed-off-by: Rob Clark <robdclark@gmail.com>