clk: tegra: dfll: Move setup code into ->probe()
Commit db7a4abcce82 ("clk: tegra: Add Tegra124 DFLL clocksource platform
driver") adds a DFLL clock driver for Tegra, but executes code in the
module init function without checking which SoC it's running on. This is
causing breakage for other SoCs in multi-platform kernels.

Fix this by moving the code into the driver's ->probe() function; there
is no reason it can't be there. Doing so also allows us to get rid of a
global variable that was used to pass data from module init to driver
probe.

Reported-by: Kevin Hilman <khilman@kernel.org>
Signed-off-by: Thierry Reding <treding@nvidia.com>