Bluetooth: Unwind l2cap_sock_shutdown()
l2cap_sock_shutdown() is designed to only action shutdown
of the channel when shutdown is not already in progress.
Therefore, reorganise the code flow by adding a goto
to jump to the end of function handling when shutdown is
already being actioned. This removes one level of code
indentation and make the code more readable.

Note that the line
sk->sk_shutdown = SHUTDOWN_MASK;
has been moved to before waiting for ACKs. This is a
different behaviour to the old code. This line will inhibit
new PDUs being created during the waiting for ACKs period
so no new I frames will be sent and hence no additional ACKs
will be need to be waited for.

Signed-off-by: Dean Jenkins <Dean_Jenkins@mentor.com>
Signed-off-by: Marcel Holtmann <marcel@holtmann.org>